UIS449: Upgraded NEST version and majority of test search interface complete

+++ src/SFA.Apprenticeships.Web.Candidate/ViewModels/VacancySearch/VacancySearchResponseViewModel.cs
@@ -4,7 +4,7 @@
 
     public class VacancySearchResponseViewModel
     {
-        public int TotalHits { get; set; }
+        public long TotalHits { get; set; }
         public int PageSize { get; set; }
         public IEnumerable<VacancySummaryViewModel> Vacancies { get; set; }
         public VacancySearchViewModel VacancySearch { get; set; }
@@ -37,7 +37,7 @@
                 var pages = 1;
                 if (PageSize > 0)
                 {
-                    pages = TotalHits/PageSize;
+                    pages = (int)TotalHits/PageSize;
                     if (TotalHits%PageSize > 0) pages++;
                 }
